Moved prebuilt classes to their own folder and namespace
