+++ src/Prebuilt/ValidateEmail.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace ValidateGetPostInput\Prebuilt;
+
+use ValidateGetPostInput\ValidateGetPostInput;
+use ValidateGetPostInput\Classes\ValidateInputSettings;
+
+/**
+ * ValidateEmail class to set the settings for the validation.
+ * This class is used to validate an email address.
+ * The email address must be a string.
+ * The email address must be at least 1 character long.
+ * The email address can be at most 320 characters long.
+ *
+ * @author  Jesse Soeteman
+ * @version 1.0
+ * @since   2023-01-18
+ */
+class ValidateEmail extends ValidateGetPostInput
+{
+    public function __construct($key, $request_type = get_input, $required = true)
+    {
+        $settings = new ValidateInputSettings();
+        $settings->input_type = $request_type;
+        $settings->pattern = validate_email_pattern;
+        $settings->required = $required;
+        $settings->isString = true;
+        $settings->min = 1;
+        $settings->max = 320;
+        parent::__construct($key, $settings);
+    }
+}
