+++ src/ValidateGetPostInput.php
@@ -155,55 +155,3 @@ class ValidateGetPostInput
         return $this->value;
     }
 }
-
-/**
- * ValidateEmail class to set the settings for the validation.
- * This class is used to validate an email address.
- * The email address must be a string.
- * The email address must be at least 1 character long.
- * The email address can be at most 320 characters long.
- *
- * @author  Jesse Soeteman
- * @version 1.0
- * @since   2023-01-18
- */
-class ValidateEmail extends ValidateGetPostInput
-{
-    public function __construct($key, $request_type = get_input, $required = true)
-    {
-        $settings = new ValidateInputSettings();
-        $settings->input_type = $request_type;
-        $settings->pattern = validate_email_pattern;
-        $settings->required = $required;
-        $settings->isString = true;
-        $settings->min = 1;
-        $settings->max = 320;
-        parent::__construct($key, $settings);
-    }
-}
-
-
-/**
- * ValidateID class to set the settings for the validation.
- * This class is used to validate an ID.
- * The ID must be a number.
- * The ID must be at least -1.
- * The ID can be at most 2147483647.
- *
- * @author  Jesse Soeteman
- * @version 1.0
- * @since   2023-01-18
- */
-class ValidateID extends ValidateGetPostInput
-{
-    public function __construct($key, $request_type = get_input, $required = true)
-    {
-        $settings = new ValidateInputSettings();
-        $settings->input_type = $request_type;
-        $settings->required = $required;
-        $settings->isString = false;
-        $settings->min = -1;
-        $settings->max = 2147483647;
-        parent::__construct($key, $settings);
-    }
-}
